encryptor-validator/requirements.txt
fastapi[all]
pycryptodome
pyqrcode
pypng
opencv